requirements.txt
# Core Django and API Framework - SECURITY UPDATED
Django>=4.2.16,<5.0  # LTS version with security patches
djangorestframework>=3.15.2,<4.0
django-cors-headers>=4.3.1

# Database
psycopg2-binary>=2.9.9

# API Documentation
drf-yasg>=1.21.7

# WSGI Server
gunicorn>=21.2.0

# Static files and utilities
whitenoise>=6.6.0
unidecode>=1.3.8

# Development and Testing
coverage>=7.3.2
factory-boy>=3.3.0
faker>=20.1.0

# Code quality
pylint>=3.0.3
pylint-django>=2.5.5
isort>=5.12.0

# HTTP requests - CVE-2023-32681 FIXED
requests>=2.31.0

# Configuration and utilities
ruamel.yaml>=0.18.5
sqlparse>=0.4.4

# PDF and Excel generation libraries
reportlab>=4.0.4
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Dotenv
python-dotenv>=1.0.0

# REMOVED DEPRECATED PACKAGES:
# unicode==2.9  # DEPRECATED - removed for security
# pytz==2021.1  # DEPRECATED - Django 4.2+ uses zoneinfo
